chatbot-agents-creator/implementation/01_foundation_layer/requirements.txt
# Core dependencies
pyyaml>=6.0
jsonschema>=4.17.3
boto3>=1.26.0
pytest>=7.3.1
pytest-cov>=4.1.0

# Development dependencies
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0